requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
spacy>=3.0.0
transformers>=4.0.0
torch>=2.1.2
scikit-learn>=0.24.0
beautifulsoup4>=4.9.0
requests>=2.25.0
numpy>=1.19.0
pandas>=1.2.0
python-dateutil>=2.8.2
aiohttp>=3.9.1
matplotlib>=3.3.0
seaborn>=0.11.0
sentence-transformers>=2.2.0
chromadb>=0.3.0 